requirements.txt
flask
langchain
langsmith
python-dotenv
openai
